+++ src/main/java/duke/TaskList.java
@@ -1,3 +1,4 @@
+package duke;
 import java.util.ArrayList;
 
 /**
