+++ src/test/java/duke/TaskListTest.java
@@ -1,3 +1,4 @@
+package duke;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
